feat: use tokens hook
